# docs-site/openapi.yaml
openapi: 3.0.3
info:
  title: SwipeSavvy Backend API
  description: |
    Multi-service backend API for SwipeSavvy Mobile Wallet — a fintech platform providing
    card issuance (FIS), payments (Authorize.Net), KYC, rewards, real-time chat, and admin management.

    ## Authentication
    Most endpoints require a JWT Bearer token obtained via `POST /api/v1/auth/login` + `POST /api/v1/auth/verify-login-otp`.

    | Token | Lifetime | Purpose |
    |-------|----------|---------|
    | Access Token | 24 hours | Authorize API requests |
    | Refresh Token | 30 days | Obtain new access tokens |

    ## Environments
    Use the server selector above to switch between Production, Staging, and Development.
  version: 1.0.0
  contact:
    name: SwipeSavvy Engineering
    email: engineering@swipesavvy.com
  license:
    name: Proprietary
servers:
  - url: http://54.224.8.14:8000
    description: Production (AWS EC2)
  - url: https://api.swipesavvy.com
    description: Production (Domain)
  - url: http://localhost:8000
    description: Development (Local)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from login flow
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT token from admin login
    WebhookSignature:
      type: apiKey
      in: header
      name: X-FIS-Signature
      description: FIS webhook signature verification
...
paths:
...
  /api/v1/notifications/register-device:
    post:
      tags: [Notifications]
      operationId: registerDevice
      summary: Register device for push notifications
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_token]
              properties:
                device_token:
                  type: string
                device_type:
                  type: string
                  default: ios
                device_name:
                  type: string
      responses:
        '200':
          description: Device registered
  /api/v1/notifications/unregister-device/{device_id}:
    post:
      tags: [Notifications]
      operationId: unregisterDevice
      summary: Unregister a device
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: device_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Device unregistered
  /api/v1/notifications/preferences:
    get:
      tags: [Notifications]
      operationId: getNotificationPreferences
      summary: Get notification preferences
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Notification preferences
    post:
      tags: [Notifications]
      operationId: updateNotificationPreferences
      summary: Update notification preferences
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                payment_notifications:
                  type: boolean
                campaign_notifications:
                  type: boolean
                support_notifications:
                  type: boolean
                security_notifications:
                  type: boolean
                feature_notifications:
                  type: boolean
      responses:
        '200':
          description: Preferences updated
  /api/v1/notifications/history:
    get:
      tags: [Notifications]
      operationId: getNotificationHistory
      summary: Get notification history
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
        - in: query
          name: notification_type
          schema:
            type: string
      responses:
        '200':
          description: Notification history
  /api/v1/notifications/mark-as-read/{notification_id}:
    post:
      tags: [Notifications]
      operationId: markNotificationRead
      summary: Mark notification as read
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: notification_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Marked as read
  /api/v1/notifications/send-event:
    post:
      tags: [Notifications]
      operationId: sendEventNotification
      summary: Send event-based notification
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_type]
              properties:
                event_type:
                  type: string
                event_data:
                  type: object
      responses:
        '200':
          description: Notification sent
  /api/v1/notifications/test:
    post:
      tags: [Notifications]
      operationId: sendTestNotification
      summary: Send test notification
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, body]
              properties:
                title:
                  type: string
                body:
                  type: string
                data:
                  type: object
      responses:
        '200':
          description: Test notification sent
  /api/v1/notifications/sns/register-device:
    post:
      tags: [Notifications]
      operationId: snsRegisterDevice
      summary: Register device with AWS SNS
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_token, device_type]
              properties:
                device_token:
                  type: string
                device_type:
                  type: string
                  enum: [ios, android]
                device_name:
                  type: string
      responses:
        '200':
          description: SNS endpoint created
  /api/v1/notifications/sns/send:
    post:
      tags: [Notifications]
      operationId: snsSendNotification
      summary: Send notification via AWS SNS
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint_arn, title, body]
              properties:
                endpoint_arn:
                  type: string
                title:
                  type: string
                body:
                  type: string
                data:
                  type: object
                badge:
                  type: integer
                sound:
                  type: string
      responses:
        '200':
          description: Notification sent
  /api/v1/notifications/sns/unregister/{endpoint_arn}:
    delete:
      tags: [Notifications]
      operationId: snsUnregisterDevice
      summary: Unregister SNS endpoint
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: endpoint_arn
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Endpoint removed
  /api/v1/notifications/sns/status/{endpoint_arn}:
    get:
      tags: [Notifications]
      operationId: snsGetEndpointStatus
      summary: Get SNS endpoint status
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: endpoint_arn
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Endpoint status

  # ═══════════════════════════════════════════
  # CHAT (10 endpoints)
  # ═══════════════════════════════════════════
...